IoT_Monitor/requirements.txt
certifi
chardet
idna
requests
six
sseclient-py
kafka-python